+++ src/main/java/forus/naviforyou/domain/findRoute/controller/FindRouteController.java
@@ -7,11 +7,11 @@ import forus.naviforyou.domain.findRoute.dto.request.WalkRouteReq;
 import forus.naviforyou.domain.findRoute.dto.response.CarRouteRes;
 import forus.naviforyou.domain.findRoute.dto.response.TravelRouteRes;
 import forus.naviforyou.domain.findRoute.dto.response.WalkRouteRes;
-import forus.naviforyou.domain.findRoute.service.FindRouteService;
+import forus.naviforyou.domain.findRoute.service.CarFindRouteService;
+import forus.naviforyou.domain.findRoute.service.TravelFindRouteService;
+import forus.naviforyou.domain.findRoute.service.WalkFindRouteService;
 import forus.naviforyou.global.common.BaseResponse;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -23,25 +23,28 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class FindRouteController {
 
-    private final FindRouteService findRouteService;
+    private final CarFindRouteService carFindRouteService;
+    private final WalkFindRouteService walkFindRouteService;
+    private final TravelFindRouteService travelFindRouteService;
+
 
     @GetMapping("/walk")
     public ResponseEntity<?> getWalkRoute(@RequestBody WalkRouteReq response){
-        WalkRouteRes walkRouteRes = findRouteService.getWalkRoute(response);
+        WalkRouteRes walkRouteRes = walkFindRouteService.getWalkRoute(response);
 
         return BaseResponse.ok(walkRouteRes);
     }
 
     @GetMapping("/travel")
     public ResponseEntity<?> getTravelRoute(@RequestBody TravelRouteReq response){
-        TravelRouteRes travelRouteReq = findRouteService.getTravelRoute(response);
+        TravelRouteRes travelRouteReq =travelFindRouteService.getTravelRoute(response);
 
         return BaseResponse.ok(travelRouteReq);
     }
 
     @GetMapping("/car")
     public ResponseEntity<?> getCarRoute(@RequestBody CarRouteReq response){
-        CarRouteRes carRouteReq = findRouteService.getCarRoute(response);
+        CarRouteRes carRouteReq =carFindRouteService.getCarRoute(response);
 
         return BaseResponse.ok(carRouteReq);
     }
